+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 dev-master (2015-06-15)
 -----------------------
 * Initial development
+
+dev-master (2015-12-17)
+-----------------------
+* Moved exception to "Exception" subdirectory
+* Added "ContainerException" class
+* "NotFoundException" extends "ContainerException"